+++ starmoney/webhooks/__init__.py
@@ -1,5 +1,13 @@
 """StarMoney SDK - Webhook Utilities"""
 
+from .events import WebhookEvent
+from .payloads import DeferredTransferSentPayload, PaymentReceivedPayload, WebhookMetadata
 from .validator import WebhookValidator
 
-__all__ = ["WebhookValidator"]
+__all__ = [
+    "WebhookValidator",
+    "WebhookEvent",
+    "WebhookMetadata",
+    "DeferredTransferSentPayload",
+    "PaymentReceivedPayload",
+]